src/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions

# Core AI + Framework
openai>=1.40.0
langchain==0.2.14
langchain-community==0.2.12

# Search + Tools
duckduckgo-search>=5.3.0

# Visualization
matplotlib>=3.8.0

ollama

